ClassRoomController에서 잘못된 Query, Command 클래스이름 수정
